==> raizen-poc-customerhealthscore/notebooks/0-afi-join-transactions.ipynb <==
import pandas as pd

# Imports
dados1 = pd.read_parquet('../data/raw/3m/transaction-01.snappy.parquet')
dados2 = pd.read_parquet('../data/raw/3m/transaction-02.snappy.parquet')
dados3 = pd.read_parquet('../data/raw/3m/transaction-03.snappy.parquet')
dados4 = pd.read_parquet('../data/raw/3m/transaction-04.snappy.parquet')

# Conferindo se os dados estão corretos
dados1.info(), dados2.info(), dados3.info(), dados4.info()

# Count para ver se estamos puxando todos os dados
dados1.count(), dados2.count()

# Concatenando todos os datasets
full_transactions = pd.concat([dados1, dados2, dados3, dados4], ignore_index=True)

dados1.shape[0] + dados2.shape[0] + dados3.shape[0] + dados4.shape[0]

full_transactions.info()

#Salvando em Parquet
full_transactions.to_parquet('full_transactions.parquet')

dados1.info()
