==> raizen-poc-customerhealthscore/notebooks/2-afi-nps-churn-palive.ipynb <==
import pandas as pd

pd.set_option('display.max_columns', None)

# Import Dataset

# Importando o dataset com as variáveis do modelo churn 
churn = pd.read_parquet('../data/raw/3m/churn.parquet')

# Importando o dataset com as variáveis do modelo p-alive
palive = pd.read_parquet('../data/raw/3m/palive.parquet')

# Importando o dataset com as variáveis de costumers e nps (feito pelo Francisco notebook1)
nps = pd.read_parquet('../data/stage/3m/customers_nps_ig.parquet')

churn['p_churn'] = churn['p_churn'].fillna(0)
churn = churn.drop_duplicates(subset='consumer_id')

#2021-11-12
#145926 - 1737544 - 709969
startdate = pd.to_datetime("2021-10-24").date()
enddate = pd.to_datetime("2021-10-31").date()
palive = palive[(palive['base_date'] > startdate) & (palive['base_date'] <= enddate)]
#palive = palive[palive['base_date'] > startdate]

nps['nps']  = nps['nps'] .map({0:0,
                                   1:0,
                                   2:0,
                                   3:0,
                                   4:0,
                                   5:0,
                                   6:0,
                                   7:0,
                                   8:1,
                                   9:1,
                                   10:1
                             },
                             na_action=None)

nps['nps'].value_counts()

palive['consumer_id'].value_counts()

nps.info()

churn.info()

palive.info()

# Merge 'nps' + 'churn'

# merge entre 'nps' e 'churn'
nps_churn = pd.merge(nps, churn, on='consumer_id', how='left')

nps_churn.info()

# Merge entre 'nps_churn' + 'palive'

# merge entre 'nps' e 'palive'
nps_palive = pd.merge(nps_churn, palive, on='consumer_id', how='left')

nps_palive.info()

columns = ['consumer_id', 'role_id', 'cpf', 'email',
       'birthdate', 'city', 'state', 'phonenumber', 'created_at', 'updated_at',
       'eou_synced_at', 'last_activity_date', 'completed_at',
       'paypal_linked_at', 'braintree_id', 'braintree_token', 'paypal_id',
       'paypal_billing_agreement', 'paypal_email', 'paypal_cpf',
       'paydiant_uri', 'paydiant_linked_at', 'stone_wallet_id',
       'stone_linked_at', 'allow_newsletter', 'allow_promos',
       'paypal_allow_automatic_payment', 'allow_automatic_payment',
       'paypal_subscription_canceled', 'active', 'signup_complete',
       'legacy_email', 'legacy_paypal_id', 'legacy_paypal_subscription',
       'city_id', 'referral_code', 'mercadopago_linked_at',
       'mercadopago_email', 'mercadopago_id', 'mfa_validated', 'country',
       'nota_recomendacao_customer', 'nota_recomendacao_nps', 'nps',
       'p_churn', 'p_currently_alive']

nps_palive = nps_palive[columns]

# Salvando

#Salvando em Parquet
nps_palive.to_parquet('../data/stage/3m/customers_nps_churn_palive.parquet')
