==> raizen-poc-customerhealthscore/notebooks/4-2-fpn-merge-nps-zendesk.ipynb <==
import pandas as pd

pd.set_option('display.max_columns', None)

# Importando o dataset completo de 'nps' (notebook5)
df_nps = pd.read_parquet('../data/stage/3m/customers_nps_churn_palive.parquet')

df_zendesk = pd.read_parquet('../data/raw/3m/zendesk-cpf.snappy.parquet')

### Checking Dataframes Info

df_nps.head()

df_zendesk.head()

df_zendesk.info()

df_nps.info()

### Converting Dtypes to merge DataFrames

df_nps['cpf'] = df_nps['cpf'].astype('int').astype('str')

df_nps['cpf'] = df_nps['cpf'].apply(lambda x: x.zfill(11)).astype('object')

### How many customer had a ticket in the last 3 months?

df_zendesk['cpf'].value_counts().shape[0]

(100*12442)/2277877

#### Somente 0.5% de clientes que transacionaram nos ultimos 3 meses abriram um ticket

### Merging DataFrames and checking results

df_customers_zendesk = pd.merge(df_nps, df_zendesk, on='cpf', how='left')

df_customers_zendesk.info()

df_customers_zendesk[df_customers_zendesk['ticket_id'].notna()]['consumer_id'].value_counts().shape

df_customers_zendesk[df_customers_zendesk['ticket_id'].notna()]['consumer_id'].value_counts()

Dentro dos clientes que responderam as pesquisa de nps ou indicaram alguem para o aplicativo a partir da data de 25-10-2021 somente 91 abriram algum tipo de chamado ao suporte, alguns abriram mais de 1 chamado, com o maximo de 6 chamados por um unico cliente, por conta disso o total de tickets abertos que constam na base totalizam 91.


### Saving DataFrame

df_customers_zendesk.to_parquet('../data/stage/3m/customers_zendesk.parquet')
