==> raizen-poc-customerhealthscore/notebooks/4-3-fpn-merge-nps-push.ipynb <==
import pandas as pd

pd.set_option('display.max_columns', None)

push_opt = pd.read_parquet('../data/raw/3m/80464_push_opt_in.snappy.parquet')
push_inbox = pd.read_parquet('../data/raw/3m/80464_push_inbox_sent.snappy.parquet')
pd.read_parquet('../data/raw/3m/80464_open.snappy.parquet') #email

push_sent = pd.read_parquet('../data/raw/3m/80464_push_sent.snappy.parquet')
push_bounced = pd.read_parquet('../data/raw/3m/80464_push_bounced.snappy.parquet')
push_opened = pd.read_parquet('../data/raw/3m/80464_push_opened.snappy.parquet')

df_nps = pd.read_parquet('../data/stage/3m/customers_nps_churn_palive.parquet')

push_sent = pd.read_parquet('../data/raw/3m/80464_push_sent.snappy.parquet')
push_sent = push_sent.drop_duplicates(subset='EVENT_UUID')

push_opened = pd.read_parquet('../data/raw/3m/80464_push_opened.snappy.parquet')
push_opened = push_opened.drop_duplicates(subset=('PUSH_ID', 'LAUNCH_ID'))

push_bounced = pd.read_parquet('../data/raw/3m/80464_push_bounced.snappy.parquet')
push_bounced = push_bounced.drop_duplicates(subset=('PUSH_ID', 'LAUNCH_ID'))

renaming_columns = {
    'EVENT_CAPTURED_DT': 'BOUNCED_EVENT_CAPTURED_DT',
    'EVENT_UUID': 'BOUNCED_EVENT_UUID'
}
push_bounced = push_bounced.rename(columns=renaming_columns)
push_bounced.loc[:,'CUSTOMER_ID'] = pd.to_numeric(push_bounced['CUSTOMER_ID'], errors='coerce', downcast='integer')
push_bounced = push_bounced[push_bounced['CUSTOMER_ID'].notna()]
push_bounced.loc[:,'CUSTOMER_ID'] = push_bounced['CUSTOMER_ID'].astype('int')
push_bounced['PUSH_BOUNCED'] = 1

columns = [
    'RIID', 'BOUNCED_EVENT_UUID','LAUNCH_ID', 'PUSH_ID',
    'CUSTOMER_ID', 'BOUNCED_EVENT_CAPTURED_DT', 'PUSH_BOUNCED'
]
push_bounced = push_bounced[columns]

renaming_columns = {
    'EVENT_CAPTURED_DT': 'OPEN_EVENT_CAPTURED_DT',
    'EVENT_UUID': 'OPEN_EVENT_UUID'
}
push_opened = push_opened.rename(columns=renaming_columns)
push_opened.loc[:,'CUSTOMER_ID'] = pd.to_numeric(push_opened['CUSTOMER_ID'], errors='coerce', downcast='integer')
push_opened = push_opened[push_opened['CUSTOMER_ID'].notna()]
push_opened.loc[:,'CUSTOMER_ID'] = push_opened['CUSTOMER_ID'].astype('int')
push_opened['PUSH_OPEN'] = 1

columns = [
    'RIID', 'OPEN_EVENT_UUID','LAUNCH_ID', 'PUSH_ID',
    'CUSTOMER_ID', 'OPEN_EVENT_CAPTURED_DT', 'PUSH_OPEN'
]
push_opened = push_opened[columns]

push_sent.loc['CUSTOMER_ID'] = pd.to_numeric(push_sent['CUSTOMER_ID'], errors='coerce', downcast='integer')

push_sent_open = push_sent.merge(push_opened, on=('PUSH_ID', 'LAUNCH_ID'), how='left', suffixes=('', '_OPEN'))

push_sent_open_bounced = push_sent_open.merge(push_bounced, on=('PUSH_ID', 'LAUNCH_ID'), how='left', suffixes=('', '_BOUNCED'))

push_sent_open_bounced.head()

df_nps.head()

nps_push_sent_open_bounced = pd.merge(df_nps, push_sent_open_bounced, left_on='consumer_id', right_on='CUSTOMER_ID', how='left', suffixes=('_nps', '_PUSHS'))

df_nps.shape

teste.shape

nps_push_sent_open_bounced.shape

#### Saving DataFrame

nps_push_sent_open_bounced.to_parquet('../data/stage/3m/customers_push.parquet')
