==> raizen-poc-customerhealthscore/notebooks/4-4-fpn-merge-nps-emails.ipynb <==
import pandas as pd

pd.set_option('display.max_columns', None)

# Importando o dataset completo de 'nps' (notebook2)
nps = pd.read_parquet('../data/stage/3m/customers_nps_churn_palive.parquet')

email_open = pd.read_parquet('../data/raw/3m/80464_open.snappy.parquet')
email_sent = pd.read_parquet('../data/raw/3m/80464_sent.snappy.parquet')

email_sent['EMAIL_SENT'] = 1
columns = [
    'CUSTOMER_ID', 'EVENT_CAPTURED_DT', 'CAMPAIGN_ID', 'LAUNCH_ID',
    'EMAIL_SENT'
]
email_sent = email_sent[columns]

email_open['EMAIL_OPEN'] = 1
columns = [
    'CUSTOMER_ID', 'EVENT_CAPTURED_DT', 'CAMPAIGN_ID', 'LAUNCH_ID',
    'EMAIL_OPEN'
]
email_open = email_open[columns]

email_open.loc[:,'CUSTOMER_ID'] = pd.to_numeric(email_open['CUSTOMER_ID'], errors='coerce', downcast='integer')

email_sent.loc[:,'CUSTOMER_ID'] = pd.to_numeric(email_sent['CUSTOMER_ID'], errors='coerce', downcast='integer')

### Merge

nps_email_open = pd.merge(nps, email_open, left_on='consumer_id', right_on='CUSTOMER_ID', how='left', suffixes=('_nps', '_EMAIL_OPEN'))

nps_email_sent = pd.merge(nps, email_sent, left_on='consumer_id', right_on='CUSTOMER_ID', how='left', suffixes=('_nps', '_EMAIL_SENT'))

### Saving DataFrame

nps_email_open.to_parquet('../data/stage/3m/customers_emails_open.parquet')

nps_email_sent.to_parquet('../data/stage/3m/customers_emails_sent.parquet')
