==> raizen-poc-customerhealthscore/notebooks/4-5-fpn-merge-nps-loyalty.ipynb <==
# Importando o pacote Pandas como 'pd'
import pandas as pd

pd.set_option('display.max_columns', None)

# Importando o dataset completo de 'nps' (notebook2)
nps = pd.read_parquet('../data/stage/3m/customers_nps_churn_palive.parquet')

df_loyalty = pd.read_parquet('../data/raw/3m/acelera_prd_loyaltytransactions.parquet')

df_loyalty.loc[:,'consumer_id'] = pd.to_numeric(df_loyalty['consumer_id'], errors='coerce', downcast='integer')

df_loyalty.head()

df_nps_loyalty = pd.merge(nps, df_loyalty, on='consumer_id', how='left', suffixes=('_nps', '_loyalty'))

df_nps_loyalty.to_parquet('../data/stage/3m/customers_loyalty.parquet')
