==> raizen-poc-customerhealthscore/notebooks/6-fpn-features-cleaning.ipynb <==
import pandas as pd
import numpy as np

pd.set_option('display.max_columns', None)

df = pd.read_csv('../data/stage/3m/features_raw.csv')

len(df.columns.tolist())

rename_columns = {
    'mob_x': 'mob',
    'sum_loyalty_discount_amount_1m_x': 'sum_loyalty_discount_amount_1m',
    'sum_loyalty_discount_amount_1m_y': 'mm_loyalty_amount',
    'distinct_voucher_id_1m_x': 'distinct_voucher_id_1m',
    'distinct_voucher_id_1m_y': 'mode_channel_3m',
    'nps': 'target'
}
df = df.rename(columns=rename_columns)

columns = [
    'consumer_id', 'target', 'mercadopago_id', 'allow_promos','total_sent_pushs_1m',
    'total_open_pushs_1m', 'total_pushs_bounced_1m','total_pushs_bounced_3m',
    'total_emails_open_1m','total_emails_open_3m','total_emails_sent_1m','total_emails_sent_3m',
    'distinct_reason_1m','distinct_reason_3m','mode_rating_reason_3m','max_diff_total_ratings',
    'mm_total_ratings_reason','mm_total_ratings','trend_total_ratings',
    'mean_rating_1m','mean_rating_3m','distinct_voucher_id_1m','mode_channel_3m',
    'mode_channel_3m','mm_vouchers','mm_redeemed_vouchers','trend_redeemed_vouchers',
    'distinct_reward_id_3m','mean_redeemed_discount_amount_3m','sum_redeemed_discount_amount_3m',
    'p_discount_1m','mean_total_value_3m','sum_total_discont_3m','max_diff_stores',
    'mm_transactions_partners','mm_total_value_partners','mob',
    'sum_loyalty_discount_amount_1m','mm_loyalty_amount', 'max_num_comments_3m',
    'tickets_1m', 'tickets_3m', 'sum_bad_rating_score_tickets'
]
df = df[columns]

df['mode_rating_reason_3m'] = df['mode_rating_reason_3m'].replace({'[]':np.nan})

df['mode_channel_3m']

df['mode_channel_3m_'] = df['mode_channel_3m'].iloc[:, 0]
df = df.drop('mode_channel_3m', axis=1)

df = df.rename(columns={'mode_channel_3m_':'mode_channel_3m'})

df['mercadopago_id'] = df['mercadopago_id'].astype(int)

interger_columns = [
    'target', 'mercadopago_id', 'allow_promos', 'total_sent_pushs_1m', 'total_open_pushs_1m',
    'total_pushs_bounced_1m', 'total_pushs_bounced_3m', 'total_emails_open_1m',
    'total_emails_open_3m', 'total_emails_sent_1m', 'total_emails_sent_3m', 'distinct_reason_1m',
    'distinct_reason_3m', 'max_diff_total_ratings', 'distinct_voucher_id_1m',
    'trend_redeemed_vouchers', 'distinct_reward_id_3m', 'max_diff_stores', 'mob',
    'max_num_comments_3m', 'tickets_1m', 'tickets_3m', 'sum_bad_rating_score_tickets'
]

float_features = [
    'mm_total_ratings_reason', 'mm_total_ratings', 'trend_total_ratings', 'mean_rating_1m',
    'mean_rating_3m', 'mm_redeemed_vouchers'
]
float_features_2 = [
    'mm_vouchers', 'mean_redeemed_discount_amount_3m',
    'sum_redeemed_discount_amount_3m', 'p_discount_1m', 'mean_total_value_3m',
    'sum_total_discont_3m', 'mm_transactions_partners'
]
float_features_3 = [
    'sum_loyalty_discount_amount_1m', 'mm_loyalty_amount'
]

#mode_rating_reason_3m
#mode_channel_3m
#'mercadopago_id', 'allow_promos'

df[interger_columns] = df[interger_columns].fillna(0).astype('int')

df[float_features] = df[float_features].fillna(0).astype('float')
df[float_features_2] = df[float_features_2].fillna(0).astype('float')
df[float_features_3] = df[float_features_3].fillna(0).astype('float')

df = df.reset_index(drop=True)

categorical_features = [
    'mode_rating_reason_3m', 'mode_channel_3m', 'mercadopago_id', 'allow_promos'
]
numerical_features = [
    'mercadopago_id', 'allow_promos', 'total_sent_pushs_1m', 'total_open_pushs_1m',
    'total_pushs_bounced_1m', 'total_pushs_bounced_3m', 'total_emails_open_1m',
    'total_emails_open_3m', 'total_emails_sent_1m', 'total_emails_sent_3m', 'distinct_reason_1m',
    'distinct_reason_3m', 'max_diff_total_ratings', 'distinct_voucher_id_1m', 'mode_channel_3m',
    'trend_redeemed_vouchers', 'distinct_reward_id_3m', 'max_diff_stores', 'mob',
    'mm_total_ratings_reason', 'mm_total_ratings', 'trend_total_ratings', 'mean_rating_1m',
    'mean_rating_3m', 'mm_redeemed_vouchers', 'mm_vouchers', 'mean_redeemed_discount_amount_3m',
    'sum_redeemed_discount_amount_3m', 'p_discount_1m', 'mean_total_value_3m',
    'sum_total_discont_3m', 'mm_transactions_partners', 'mm_vouchers',
    'mean_redeemed_discount_amount_3m', 'sum_redeemed_discount_amount_3m',
    'p_discount_1m', 'mean_total_value_3m', 'sum_total_discont_3m', 'mm_transactions_partners',
    'sum_loyalty_discount_amount_1m', 'mm_loyalty_amount', 'max_num_comments_3m',
    'tickets_1m', 'tickets_3m', 'sum_bad_rating_score_tickets'
]
target_column = 'target'

### Saving DataFrame

#Salvando em Parquet
df.to_parquet('../data/stage/3m/features_gold.parquet')
